+++ CoreArrangementPlugins/src/au/gov/asd/tac/constellation/plugins/arrangements/SelectedInclusionGraph.java
@@ -18,7 +18,6 @@ package au.gov.asd.tac.constellation.plugins.arrangements;
 import au.gov.asd.tac.constellation.graph.Graph;
 import au.gov.asd.tac.constellation.graph.GraphElementType;
 import au.gov.asd.tac.constellation.graph.GraphWriteMethods;
-import au.gov.asd.tac.constellation.graph.LayersConcept;
 import au.gov.asd.tac.constellation.graph.schema.visual.concept.VisualConcept;
 
 /**